+++ firmware/battery-tester.X/menu_navigation.c
@@ -1250,10 +1250,10 @@ void Menu(BattParameters *bat_param)
                     
                     if(Button_EventGet(RIGHT_SW))
                     {
-                        bat_param->charge_current_4_percent=bat_param->charge_current_4_percent+val;
+                        bat_param->charge_current_4_percent = bat_param->charge_current_4_percent+(uint8_t)val;
                     }else if(Button_EventGet(LEFT_SW))
                     {
-                        bat_param->charge_current_4_percent=bat_param->charge_current_4_percent-val;
+                        bat_param->charge_current_4_percent = bat_param->charge_current_4_percent-(uint8_t)val;
                     }
 
                     if(bat_param->charge_current_4_percent<MIN_PRECENT_CHARGE_VAL)
